service/generate_api_credentials.py
from apiclient.discovery import build
from httplib2 import Http
from oauth2client import file, client, tools

SCOPES = 'https://www.googleapis.com/auth/spreadsheets https://www.googleapis.com/auth/gmail.send'
store = file.Storage('credentials.json')
creds = store.get()
if not creds or creds.invalid:
    flow = client.flow_from_clientsecrets('client_secret.json', SCOPES)
    creds = tools.run_flow(flow, store)

email_service = build('gmail', 'v1', http=creds.authorize(Http()))
spreadsheet_service = build('sheets', 'v4', http=creds.authorize(Http()))
